app/src/test/java/org/clintonhealthaccess/lmis/app/services/CommodityServiceTest.java
package org.clintonhealthaccess.lmis.app.services;

import com.google.inject.Inject;

import org.clintonhealthaccess.lmis.app.persistence.CommoditiesRepository;
import org.clintonhealthaccess.lmis.utils.RobolectricGradleTestRunner;
import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;

import static org.clintonhealthaccess.lmis.utils.TestInjectionUtil.setUpInjection;
import static org.hamcrest.number.OrderingComparison.greaterThan;
import static org.junit.Assert.assertThat;

@RunWith(RobolectricGradleTestRunner.class)
public class CommodityServiceTest {
    @Inject
    private CommodityService commodityService;

    @Inject
    private CommoditiesRepository commoditiesRepository;

    @Before
    public void setUp() throws Exception {
        setUpInjection(this);
    }

    @Test
    public void testShouldPrepareDefaultCommodities() throws Exception {
        commodityService.initialise();

        assertThat(commoditiesRepository.allCategories().size(), greaterThan(0));
    }
}